+++ src/Exception/FormatException.php
@@ -1,0 +1,23 @@
+<?php
+/**
+ * This file is part of the "litgroup/datetime" package.
+ *
+ * (c) Roman Shamritskiy <roman@litgroup.ru>
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
+declare(strict_types = 1);
+
+namespace LitGroup\Time\Exception;
+
+use Exception;
+
+class FormatException extends Exception
+{
+    public function __construct($message)
+    {
+        parent::__construct($message);
+    }
+}